\name{aconfrs}

\alias{aconfrs}

\title{
	aconfrs - Academic Conference Management System
}

\description{
	Easy and simple for small university congresses.
	
	\enumerate{	
		\item In development phase.
		\item Not yet fully functional.
	}
}

\usage{ aconfrs() }

\arguments{ No arguments }

\value{ No return value. Interactive Session of Shiny App }

\details{
	A Shiny App for managing University Congresses
}

\references{
	\url{https://github.com/cleber-n-borges/aconfrs}
}

\examples{ #aconfrs() }

\concept{CoMS}
